fix(provider-aiven): use 64-bit types for partition offsets and size
PartitionInfo declared earliest_offset, latest_offset and size as
Integer, but Aiven reports size in bytes and Kafka offsets are 64-bit.
With no lenient numeric coercion on the mapper, a partition over 2GiB
aborted deserialization instead of overflowing silently:

  JsonMappingException: Numeric value (3221225472) out of range of int

Since AivenKafkaTopicCollector fetches topic info per topic, one
oversized partition failed the entire topic listing and diff path.

Widen the three fields to Long; the partition index stays Integer as
Kafka partition ids are int32. Add KafkaTopicAdapterTest covering an
oversized size, oversized offsets, and the V1KafkaTopic mapping.

+++ providers/jikkou-provider-aiven/src/main/java/io/jikkou/extension/aiven/api/data/PartitionInfo.java
@@ -9,11 +9,22 @@ package io.jikkou.extension.aiven.api.data;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.jikkou.core.annotation.Reflectable;
 
+/**
+ * Information about a single topic partition.
+ *
+ * <p>Offsets and size are 64-bit: Kafka offsets are {@code long}, and the partition size is reported in bytes,
+ * so it exceeds the 32-bit range as soon as a partition grows past 2GiB.
+ *
+ * @param earliestOffset the earliest offset available in the partition.
+ * @param latestOffset   the latest offset available in the partition.
+ * @param partition      the partition index.
+ * @param size           the partition size in bytes.
+ */
 @Reflectable
 public record PartitionInfo(
-    @JsonProperty("earliest_offset") Integer earliestOffset,
-    @JsonProperty("latest_offset") Integer latestOffset,
+    @JsonProperty("earliest_offset") Long earliestOffset,
+    @JsonProperty("latest_offset") Long latestOffset,
     @JsonProperty("partition") Integer partition,
-    @JsonProperty("size") Integer size
+    @JsonProperty("size") Long size
 ) {
 }
